requirements.txt
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
asgiref==3.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
dash==3.2.0
dash-bootstrap-components==2.0.4
dash_auth==2.3.0
dill==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
doclayout_yolo==0.0.4
drill==1.2.0
et_xmlfile==2.0.0
filelock==3.13.1
Flask==3.1.2
fonttools==4.60.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
narwhals==2.6.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.3.1
polars==1.33.1
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.4
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
retrying==1.4.2
rpds-py==0.27.1
safetensors==0.6.2
scipy==1.16.2
seaborn==0.13.2
setuptools==70.2.0
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.50.0
stringzilla==4.0.15
sympy==1.13.3
tenacity==9.1.2
thop==0.1.1.post2209072238
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
xlsxwriter==3.2.9
zipp==3.23.0
